initial setup for typescript
