--- src/index.js
@@ -1,6 +1,0 @@
-import ARToolkitNFT from './ARToolkitNFT'
-import ARControllerNFT from './ARControllerNFT'
-
-export default {
-  ARToolkitNFT, ARControllerNFT
-}
