+++ src/index.ts
@@ -1,0 +1,6 @@
+import ARToolkitNFT from './ARToolkitNFT'
+import ARControllerNFT from './ARControllerNFT'
+
+export default {
+  ARToolkitNFT, ARControllerNFT
+}
